% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fonts.R
\name{check_fonts_in_r}
\alias{check_fonts_in_r}
\title{Tests if fonts are available on the system and import for extrafont}
\usage{
check_fonts_in_r(fonts = c("Raleway", "Roboto Slab", "Nanum Pen"),
  import = TRUE, verbose = TRUE, browse = FALSE)
}
\arguments{
\item{fonts}{vector of fonts}

\item{import}{Logical. Whether to import ttf in R. (Otherwise just check)}

\item{verbose}{list fonts to install if needed}

\item{browse}{open directory with fonts to install, when verbose is TRUE}
}
\description{
Tests if fonts are available on the system and import for extrafont
}
\details{
For fonts to be used, they need to be installed on the machine.
}
\examples{
# Change browse to TRUE to open folder with fonts required
check_fonts_in_r(import = FALSE)

}
