% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{dreal_cols}
\alias{dreal_cols}
\title{Function to extract dreal colors as hex codes}
\usage{
dreal_cols(...)
}
\arguments{
\item{...}{Character names of dreal_colors}
}
\description{
Function to extract dreal colors as hex codes
}
\examples{
dreal_cols("primary", "danger")
}
