% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-theme.R
\name{dreal_geom_defaults}
\alias{dreal_geom_defaults}
\title{Change geom defaults to the dreal colors}
\usage{
dreal_geom_defaults(default = c("dreal", "ggplot2"))
}
\arguments{
\item{default}{Choose which default to set "dreal", "ggplot2"}
}
\description{
Change geom defaults to the dreal colors
}
