% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{dreal_pal}
\alias{dreal_pal}
\title{Return function to interpolate a dreal color palette}
\usage{
dreal_pal(palette = "continuous", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in dreal_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a dreal color palette
}
\examples{
dreal_pal("continuous")(6)
}
