% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-theme.R
\name{drealize}
\alias{drealize}
\title{transform a ggplot with the DREAL theme}
\usage{
drealize(g, legend.position = "bottom", title = "Title of the graph",
  caption = "Source : DREAL", heights, widths, ...)
}
\arguments{
\item{g}{a ggplot}

\item{legend.position}{position of the legend "bottom" or "right"}

\item{title}{title of the graph}

\item{caption}{cpation of the graph}

\item{heights}{heights of the different parts of the plot.
Default to c(1, 9, 1, 0.5) for legend.position = "bottom".
Default to c(1, 9, 0.5) for legend.position = "left".}

\item{widths}{widths of the two columns of plot with legend.position = "left".
Default to c(4, 1).}

\item{...}{other parameters of \code{ggplot2::\link[ggplot2]{theme}}}
}
\description{
transform a ggplot with the DREAL theme
}
\examples{
library(ggplot2)
g <- ggplot(mtcars, aes(hp, mpg, colour = as.character(gear))) +
geom_point(
size = 4, alpha = .8) +
scale_color_dreal_d()

g2 <- drealize(g,
title = "Plot created with drealize()",
caption = "Source: DREAL",
legend.position = "bottom")

}
