% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{drealRampPaletteDiscrete}
\alias{drealRampPaletteDiscrete}
\title{drealRampPaletteDiscrete}
\usage{
drealRampPaletteDiscrete(colors)
}
\arguments{
\item{colors}{colors to interpolate; must be a valid argument to col2rgb().}
}
\description{
drealRampPaletteDiscrete
}
