% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drealthemes-package.R
\docType{package}
\name{drealthemes-package}
\alias{drealthemes}
\alias{drealthemes-package}
\title{drealthemes: Themes for DREAL plots}
\description{
Color palettes and ggplot2 themes.
}
\author{
\strong{Maintainer}: Sebastien Rochette \email{sebastien@thinkr.fr} (0000-0002-1565-9313)

Authors:
\itemize{
  \item Colin Fay \email{colin@thinkr.fr} (0000-0001-7343-1846)
}

Other contributors:
\itemize{
  \item DREAL [copyright holder]
}

}
\keyword{internal}
