% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-theme.R
\name{ggempty}
\alias{ggempty}
\title{Create empty ggplot on which to add legend}
\usage{
ggempty(plot = NULL, xlim = c(0, 1), ylim = c(0, 1), ...)
}
\arguments{
\item{plot}{The plot to use as a starting point. Can be a ggplot2 plot, an arbitrary
grob or gtable, or a recorded base-R plot, as in \code{\link[=plot_to_gtable]{plot_to_gtable()}}.}

\item{xlim}{The x-axis limits for the drawing layer.}

\item{ylim}{The y-axis limits for the drawing layer.}

\item{...}{other parameters of \code{ggplot2::\link[ggplot2]{theme}}}
}
\description{
Create empty ggplot on which to add legend
}
