% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{list_dreal_pals}
\alias{list_dreal_pals}
\title{List dreal palettes available with recommended type}
\usage{
list_dreal_pals()
}
\description{
List dreal palettes available with recommended type
}
\examples{
list_dreal_pals()
}
