% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scales.R
\name{scale_color_dreal_c}
\alias{scale_color_dreal_c}
\title{Continuous color scale constructor for dreal colors}
\usage{
scale_color_dreal_c(..., palette = "continuous", reverse = FALSE)
}
\arguments{
\item{...}{Additional arguments passed to scale_color_gradientn()}

\item{palette}{Character name of palette in dreal_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}
}
\description{
Continuous color scale constructor for dreal colors
}
