% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-functions.R
\name{scale_fill_dreal_c}
\alias{scale_fill_dreal_c}
\title{Continuous fill scale constructor for dreal colors}
\usage{
scale_fill_dreal_c(palette = "continuous", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in dreal_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to scale_fill_gradientn()}
}
\description{
Continuous fill scale constructor for dreal colors
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
   geom_point(size = 4) +
   scale_color_dreal_c()
}
