% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{show_pal}
\alias{show_pal}
\title{Show palette of colors}
\usage{
show_pal(cols, cols_names, nrow = 1, ncol = length(cols),
  show_text = TRUE)
}
\arguments{
\item{cols}{vector of colors. Can be named vector.}

\item{cols_names}{names of colors if cols not named vector}

\item{nrow}{number of rows of the graph output}

\item{ncol}{number of cols of the graph output}

\item{show_text}{Logical. Whether to show cols_names}
}
\description{
Show palette of colors
}
\examples{
show_pal(cols = dreal_pal("continuous")(6), show_text = FALSE)
}
