% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-theme.R
\name{theme_dreal_dark}
\alias{theme_dreal_dark}
\alias{theme_dreal_light}
\alias{theme_dreal}
\title{ggplot2 themes for Dreal}
\usage{
theme_dreal_dark(legend.position = c("bottom", "right"),
  caption.position = c("left", "right"), ...)

theme_dreal_light(legend.position = c("bottom", "right"),
  caption.position = c("left", "right"), ...)

theme_dreal(type = c("light", "dark"), legend.position = c("bottom",
  "right"), caption.position = c("left", "right"), ...)
}
\arguments{
\item{legend.position}{position of the legend "bottom" or "right"}

\item{caption.position}{position of the caption "left" or "right"}

\item{...}{Other parameters of \code{\link[ggplot2]{theme}}}

\item{type}{Choose theme type}
}
\description{
These are a complete themes which control all non-data display. Use theme() if you just need to tweak the display of an existing theme
}
