% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fonts.R
\name{check_fonts_in_r}
\alias{check_fonts_in_r}
\title{Tests if fonts are available on the system and import for extrafont}
\usage{
check_fonts_in_r(fonts = c("Raleway", "Roboto Slab", "Caveat"),
  import = TRUE, verbose = TRUE, browse = FALSE)
}
\arguments{
\item{fonts}{vector of fonts}

\item{import}{Logical. Whether to import ttf in R. (Otherwise just check)}

\item{verbose}{list fonts to install if needed}

\item{browse}{open directory with fonts to install, when verbose is TRUE}
}
\description{
Tests if fonts are available on the system and import for extrafont
}
\details{
For fonts to be used, they need to be installed on the machine.
}
\examples{
# Change browse to TRUE to open folder with fonts required
check_fonts_in_r(import = FALSE)

# Test if working
\dontrun{
library(ggplot2)
ggplot(mtcars) +
  geom_point(aes(cyl, mpg)) +
  labs(title = "Mon titre avec Roboto Slab",
       x = "lab X avec Raleway",
       y = "Lab Y avec Caveat") +
  theme(title = element_text(family = "Roboto Slab"),
        axis.title.x = element_text(family = "Raleway"),
        axis.title.y = element_text(family = "Caveat")
  )
}
}
