% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scales.R
\name{scale_color_dreal_d}
\alias{scale_color_dreal_d}
\alias{scale_colour_dreal_d}
\title{Discrete color scale constructor for dreal colors}
\usage{
scale_color_dreal_d(..., palette = "discrete_long", reverse = FALSE)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::\link[ggplot2]{discrete_scale}}}

\item{palette}{Character name of palette in dreal_palettes or "explore" for high number of levels}

\item{reverse}{Boolean indicating whether the palette should be reversed}
}
\description{
Discrete color scale constructor for dreal colors
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
   geom_point(size = 4) +
   scale_color_dreal_d()
# Long scales in exploration phase
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = as.factor(1:nrow(iris)))) +
  geom_point(size = 4) +
  scale_color_dreal_d(palette = "explore", guide = FALSE)
}
