% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-functions.R
\name{scale_color_dreal_d}
\alias{scale_color_dreal_d}
\title{Discrete color scale constructor for dreal colors}
\usage{
scale_color_dreal_d(palette = "discrete", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in dreal_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to scale_color_manual()}
}
\description{
Discrete color scale constructor for dreal colors
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
   geom_point(size = 4) +
   scale_color_dreal_d()
}
