% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-functions.R
\name{scale_fill_dreal_d}
\alias{scale_fill_dreal_d}
\title{Discrete fill scale constructor for dreal colors}
\usage{
scale_fill_dreal_d(palette = "discrete", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in dreal_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to scale_fill_manual()}
}
\description{
Discrete fill scale constructor for dreal colors
}
\examples{
ggplot(iris, aes(Sepal.Width, fill = Species)) +
  geom_histogram(size = 4, bins = 20) +
  scale_fill_dreal_d()
}
