% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-theme.R
\name{theme_dreal}
\alias{theme_dreal}
\title{ggplot2 theme for Dreal}
\usage{
theme_dreal(legend.position = "bottom", ...)
}
\arguments{
\item{legend.position}{position of the legend "bottom" or "right"}

\item{...}{Other parameters of \code{\link[ggplot2]{theme}}}
}
\description{
This is a complete theme which control all non-data display. Use theme() if you just need to tweak the display of an existing theme
}
